# shared/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BéninExpense AI - Plateforme de gestion des dépenses et conformité fiscale
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts:
    get:
      operationId: getAccounts
      tags: [accounts]
      summary: List accounts
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get account by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Not found

    patch:
      operationId: updateAccount
      tags: [accounts]
      summary: Update account balance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"

  # Expenses
...
  /ai/parse-text:
    post:
      operationId: parseExpenseText
      tags: [ai]
      summary: Parse natural language text to extract expense data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParseTextInput"
      responses:
        "200":
          description: Parsed expense data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParsedExpense"
  /ai/parse-receipt:
    post:
      operationId: parseReceipt
      tags: [ai]
      summary: OCR a receipt image and extract expense data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParseReceiptInput"
      responses:
        "200":
          description: Extracted expense data from receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParsedExpense"

  # DGI / e-MECeF
...
components:
  schemas:
...
    Account:
      type: object
      required: [id, name, type, balance, currency]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [cash, bank, mobile_money]
        balance:
          type: number
        currency:
          type: string
          default: FCFA
        createdAt:
          type: string
          format: date-time
    AccountInput:
      type: object
      required: [name, type, balance]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [cash, bank, mobile_money]
        balance:
          type: number
        currency:
          type: string
          default: XOF
          description: ISO 4217 currency code (XOF, XAF, GHS, NGN, etc.)
    AccountUpdate:
      type: object
      properties:
        name:
          type: string
        balance:
          type: number
        currency:
          type: string
          description: ISO 4217 currency code

    # Expense schemas
...
    ParseTextInput:
      type: object
      required: [text]
      properties:
        text:
          type: string
          description: Natural language text describing the expense
    ParseReceiptInput:
      type: object
      required: [imageBase64]
      properties:
        imageBase64:
          type: string
          description: Base64-encoded image of the receipt
    ParsedExpense:
      type: object
      required: [description, amount, category, confidence]
      properties:
        description:
          type: string
        amount:
          type: number
        category:
          type: string
        confidence:
          type: number
          description: Confidence score 0-1
        rawText:
          type: ["string", "null"]

    # DGI schemas
...